move type registries
 - we can export the macros types directly from index.ts because it only exists for types anyway
  - ensure-safe-component doesn't come from macros, it belongs in its own registry in embroider/util
